Completing of problem solution
Created a model class `CarModel`, containing 3 attributes (`ownerName`, `modelName` and `series`) and a method for getting a list of all `CarModel` records (currently this method returns a "baked" list of 5 `CarModel` instances).
Created a service class `CarService`, containing a method `getCars()`, which returns a list of `CarModel` records, the size of which is limited by the call parameter `count`.
Created an **HTML** template `cars.html`, based on which a table representation of the `CarModel` record list is formed.
Created a controller class `CarsController`, which is mapped to `/cars` and is responsible for forming an **HTML** page based on the `cars.html` template and the list obtained from method `CarService.getCars()`.